requirements.txt
astroid==1.4.4
behave==1.2.5
colorama==0.3.7
defusedxml==0.4.1
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.9.4
django-allauth==0.24.1
django-debug-toolbar==1.4
django-filter==0.13.0
django-localflavor==1.2
django-nose==1.4.3
django-tinymce==2.3.0
django-toolbelt==0.0.1
djangorestframework==3.3.2
factory-boy==2.6.1
fake-factory==0.5.7
gunicorn==19.4.5
lazy-object-proxy==1.2.1
Markdown==2.6.5
nose==1.3.7
numpy==1.10.4
oauthlib==1.0.3
parse==1.6.6
parse-type==0.3.4
psycopg2==2.6.1
pylint==1.5.4
pylint-django==0.7.1
pylint-plugin-utils==0.2.3
python-dateutil==2.5.1
python3-openid==3.0.9
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
sqlparse==0.1.19
static3==0.6.1
whitenoise==2.0.6
wrapt==1.10.6
